refactor(common)!: move ParquetAdapter next to its stats consumers
Relocates org.apache.hudi.ParquetAdapter to org.apache.hudi.stats,
the package of its only two consumers (ValueType, ValueMetadata). The
interface is a parquet-version compatibility shim serving column-stats
typing exclusively; its two implementations remain in
org.apache.parquet.schema (deliberate namespace residency for parquet
API access) and are loaded reflectively via class literals, so they
follow the move automatically.

BREAKING CHANGE: source/binary break for external code importing this
interface - unlikely in practice, since the static getAdapter() factory
hard-wires the only two shipped implementations. Fails loudly at class
load if it exists anywhere; never referenced by config or reflection
by name.
