+++ hudi-common/src/main/java/org/apache/hudi/stats/ValueMetadata.java
@@ -19,7 +19,6 @@
 
 package org.apache.hudi.stats;
 
-import org.apache.hudi.ParquetAdapter;
 import org.apache.hudi.avro.model.HoodieValueTypeInfo;
 import org.apache.hudi.common.schema.HoodieSchema;
 import org.apache.hudi.common.util.collection.Pair;
