+++ hudi-common/src/main/java/org/apache/parquet/schema/OriginalTypeParquetAdapter.java
@@ -19,7 +19,7 @@
 
 package org.apache.parquet.schema;
 
-import org.apache.hudi.ParquetAdapter;
+import org.apache.hudi.stats.ParquetAdapter;
 import org.apache.hudi.stats.ValueType;
 
 /**
